+++ packages/core/src/image-manipulation/effects.js
@@ -4,68 +4,6 @@ import * as constants from '../constants';
 import * as compositeModes from './composite-modes';
 
 /**
- * Applies a true Gaussian blur to the image (warning: this is VERY slow)
- * @param {number} r the pixel radius of the blur
- * @param {function(Error, Jimp)} cb (optional) a callback for when complete
- * @returns {Jimp} this for chaining of methods
- */
-export function gaussian(r, cb) {
-    // http://blog.ivank.net/fastest-gaussian-blur.html
-    if (typeof r !== 'number') {
-        return throwError.call(this, 'r must be a number', cb);
-    }
-
-    if (r < 1) {
-        return throwError.call(this, 'r must be greater than 0', cb);
-    }
-
-    const rs = Math.ceil(r * 2.57); // significant radius
-
-    for (let y = 0; y < this.bitmap.height; y++) {
-        for (let x = 0; x < this.bitmap.width; x++) {
-            let red = 0;
-            let green = 0;
-            let blue = 0;
-            let alpha = 0;
-            let wsum = 0;
-
-            for (let iy = y - rs; iy < y + rs + 1; iy++) {
-                for (let ix = x - rs; ix < x + rs + 1; ix++) {
-                    const x1 = Math.min(this.bitmap.width - 1, Math.max(0, ix));
-                    const y1 = Math.min(
-                        this.bitmap.height - 1,
-                        Math.max(0, iy)
-                    );
-                    const dsq = (ix - x) * (ix - x) + (iy - y) * (iy - y);
-                    const wght =
-                        Math.exp(-dsq / (2 * r * r)) / (Math.PI * 2 * r * r);
-                    const idx = (y1 * this.bitmap.width + x1) << 2;
-
-                    red += this.bitmap.data[idx] * wght;
-                    green += this.bitmap.data[idx + 1] * wght;
-                    blue += this.bitmap.data[idx + 2] * wght;
-                    alpha += this.bitmap.data[idx + 3] * wght;
-                    wsum += wght;
-                }
-
-                const idx = (y * this.bitmap.width + x) << 2;
-
-                this.bitmap.data[idx] = Math.round(red / wsum);
-                this.bitmap.data[idx + 1] = Math.round(green / wsum);
-                this.bitmap.data[idx + 2] = Math.round(blue / wsum);
-                this.bitmap.data[idx + 3] = Math.round(alpha / wsum);
-            }
-        }
-    }
-
-    if (isNodePattern(cb)) {
-        cb.call(this, null, this);
-    }
-
-    return this;
-}
-
-/**
  * Composites a source image over to this image respecting alpha channels
  * @param {Jimp} src the source Jimp instance
  * @param {number} x the x position to blit the image
