+++ packages/plugin-gaussian/README.md
@@ -1,0 +1,3 @@
+# @jimp/plugin-gaussian
+
+gaussian blur an image.
